Section 23 video 48 end of it Finaliser les methodes de suppression

+++ src/main/java/com/bouali/gestiondestock/services/impl/CategoryServiceImpl.java
@@ -1,10 +1,12 @@
 package com.bouali.gestiondestock.services.impl;
 
+import com.bouali.gestiondestock.Repository.ArticleRepository;
 import com.bouali.gestiondestock.Repository.CategoryRepository;
 import com.bouali.gestiondestock.dto.CategoryDto;
 import com.bouali.gestiondestock.exception.EntityNotFoundException;
 import com.bouali.gestiondestock.exception.ErrorCodes;
 import com.bouali.gestiondestock.exception.InvalidEntityException;
+import com.bouali.gestiondestock.exception.InvalidOperationException;
 import com.bouali.gestiondestock.model.Article;
 import com.bouali.gestiondestock.model.Category;
 import com.bouali.gestiondestock.services.CategoryService;
@@ -25,10 +27,12 @@ import java.util.stream.Collectors;
 public class CategoryServiceImpl implements CategoryService {
 
     private CategoryRepository categoryRepository ;
+    private ArticleRepository articleRepository ;
 
     @Autowired //injection des dependace par constructeur
-    public CategoryServiceImpl(CategoryRepository categoryRepository){
+    public CategoryServiceImpl(CategoryRepository categoryRepository,ArticleRepository articleRepository){
         this.categoryRepository=categoryRepository;
+        this.articleRepository=articleRepository;
     }
 
     @Override
@@ -90,6 +94,11 @@ public class CategoryServiceImpl implements CategoryService {
         if(id==null){
             log.error("Article ID is null") ;
             return;
+        }
+        List<Article> articles = articleRepository.findAllByCategoryId(id);
+        if (!articles.isEmpty()){
+            throw new InvalidOperationException("Impossible de supprimer cette categorie deja utiliser  ",
+                    ErrorCodes.CATEGORY_ALREADY_IN_USE) ;
         }
         categoryRepository.deleteById(id);
     }
