+++ src/main/java/com/bouali/gestiondestock/services/impl/FournisseurServiceImpl.java
@@ -1,5 +1,7 @@
 package com.bouali.gestiondestock.services.impl;
 
+import com.bouali.gestiondestock.Repository.CommandeClientRepository;
+import com.bouali.gestiondestock.Repository.CommandeFournisseurRepository;
 import com.bouali.gestiondestock.Repository.EntrepriseRepository;
 import com.bouali.gestiondestock.Repository.FournisseurRepository;
 import com.bouali.gestiondestock.dto.EntrepriseDto;
@@ -7,6 +9,9 @@ import com.bouali.gestiondestock.dto.FournisseurDto;
 import com.bouali.gestiondestock.exception.EntityNotFoundException;
 import com.bouali.gestiondestock.exception.ErrorCodes;
 import com.bouali.gestiondestock.exception.InvalidEntityException;
+import com.bouali.gestiondestock.exception.InvalidOperationException;
+import com.bouali.gestiondestock.model.CommandeClient;
+import com.bouali.gestiondestock.model.CommandeFournisseur;
 import com.bouali.gestiondestock.model.Entreprise;
 import com.bouali.gestiondestock.model.Fournisseur;
 import com.bouali.gestiondestock.services.FournisseurService;
@@ -26,10 +31,12 @@ public class FournisseurServiceImpl implements FournisseurService {
 
 
     private FournisseurRepository fournisseurRepository ;
+    private CommandeFournisseurRepository commandeFournisseurRepository ;
 
     @Autowired //injection des dependace par constructeur
-    public FournisseurServiceImpl(FournisseurRepository fournisseurRepository){
+    public FournisseurServiceImpl(FournisseurRepository fournisseurRepository,CommandeFournisseurRepository commandeFournisseurRepository){
         this.fournisseurRepository=fournisseurRepository;
+        this.commandeFournisseurRepository=commandeFournisseurRepository;
     }
 
 
@@ -73,6 +80,11 @@ public class FournisseurServiceImpl implements FournisseurService {
         if(id==null){
             log.error("Fournisseur ID is null") ;
             return;
+        }
+        List<CommandeFournisseur> commandeFournisseurs = commandeFournisseurRepository.findAllByFournisseurId(id) ;
+        if (!commandeFournisseurs.isEmpty()){
+            throw new InvalidOperationException("Impossible de supprimer un fournisseur  qui a  deja des commandes  ",
+                    ErrorCodes.FOURNISSEUR_ALREADY_IN_USE) ;
         }
         fournisseurRepository.deleteById(id);
     }
